requirements.txt
python-telegram-bot>=21.0
openai>=1.40.0
google-api-python-client>=2.128.0
google-auth>=2.32.0
google-auth-oauthlib>=1.2.0
python-dateutil>=2.9.0
PyYAML>=6.0.0
